fix: handle block elements like div within lists
Non-p block elements within lists were creating invalid nodes and simply
not appearing in the docx file. This catches some of the cases and
converts them to a plain text string. Formatting may be lost, but the
text content will render.

Example of html that would create invalid nodes:
```html
<ul>
  <li><strong>This will show in the docx</strong></li>
  <li><div>This would not show in the docx</div></li>
</ul>
```